Stock2/requirements.txt
Flask==3.0.3
numpy==1.24.4
pandas==2.0.3
tensorflow==2.8.0
tensorboard==2.8.0
tensorflow-gpu==2.8.0
tensorflow-hub==0.16.1
plotly>=5.0
finance-datareader==0.9.90
keras==2.8.0
finance-datareader==0.9.90
pandas-datareader==0.10.0
yfinance==0.2.37
matplotlib==3.7.5
matplotlib-inline==0.1.6
plotly==5.20.0
tensorflow-cpu
protobuf==3.20.0
Werkzeug==3.0.1
pillow==10.4.0